Add zephyr_whatsnew.py
This is a fairly foundries.io-specific script we use to maintain our
downstream. It has some nice features that would be good to generalize
for generating mergeup commit messages in an arbitrary
downstream. It would be nice to handle that by converting this into
one or more west extension commands.

Signed-off-by: Marti Bolivar <marti@foundries.io>